Create basic voting system
Create a basic voting system with a vote and results page.  Implement
the required models, migrations, views, controllers and requests.

The person:create command allows the creation of a person to vote for.
The vote page then allows that person to be voted for.  The results page
displays a table and pie chart of the votes.